[important] - refactor so everythign is in one place
